backend: gl: split code into multiple files
gl_common.c is getting too big.
